Add files via upload
Hello, This is a project about world layoffs in which I tried to clean the dataset as best as possible.
Giving you a quick overview First, I found duplicates that exist and removed them, Next, I standardized the data, and after that I focused on removing null values and any blank spaces in the data. and lastly, remove unnecessary columns from the data.